XeTexFont -> PlatformFontRef; xetex_layout_interface moved into xetex_layout_engine
