+++ engine/src/xetex_xetex0.rs
@@ -74,7 +74,6 @@ use crate::xetex_ini::{
 use crate::xetex_ini::{b16x4, b32x2, memory_word, prefixed_command};
 use crate::xetex_io::{input_line, open_or_close_in, set_input_file_encoding, u_close};
 use crate::xetex_layout_engine::*;
-use crate::xetex_layout_interface::*;
 use crate::xetex_opentype_math::get_ot_math_constant;
 use crate::xetex_linebreak::line_break;
 use crate::xetex_math::{
